fix mst sink resolver operator
Drop const for flow operators and their apply method. Allow it to create
or update some internal objects/state while executing it. This is
needed, e.g., for mst sink resolver which creates (and cache) a basin
graph. The design could certainly be cleaner but I think it is ok for
now.