flow graph: remove separate inline impl header
It wasn't working well with LSP servers, at least on my local config.
Alternatively I could have updated the CMake configuration or the pixi
compile commands tasks but the separate header structure was currently
only used for flow graph and the implementation code is not too
big (yet) to put in the same header file.