Flow kernel improvements (#185)
* flow graph: remove separate inline impl header

It wasn't working well with LSP servers, at least on my local config.
Alternatively I could have updated the CMake configuration or the pixi
compile commands tasks but the separate header structure was currently
only used for flow graph and the implementation code is not too
big (yet) to put in the same header file.

* fix compilation warnings

* pixi update

* typing fixes

* add support for DEPTH_DOWNSTREAM direction (sequential only)

* support setting output values at receivers

* add simple kernel drainage area test

* add support for BREATH_DOWNSTREAM direction (sequential only)

* save generated code instead of printing it

Access it via kernel object attribute.

* clean-up

* fix mypy

* allow skip data get/set access at receivers

* add some docstrings

* max_receivers default to None

* replace max_receivers by auto_resize

It is safer to set a fixed size for the receiver temporary containers
from the properties of the flow graph than from user input.

* wip access data at donors

* fix mypy

* allow accessing kernel data at flow donors

* more tests